TMemo: fix streaming field size mismatch
This is uint since 7553c77482ddd2ed4ecb3ebf2bd4f8156a91e723.

+++ source/tvision/tmemo.cpp
@@ -81,7 +81,7 @@ void TMemo::write( opstream& os )
 void *TMemo::read( ipstream& is )
 {
     TEditor::read( is );
-    ushort length;
+    uint length;
     is >> length;
     if( isValid )
         {
